For the Stubs in the Persistence  > Did some structured classes with simple interfaces  > Using ArrayList implemented as List  > no DAO but something close enough I guess  > some changes needs to be made to work with Objects
Did a bit of changing with the Objects but needs more fixing.
 > some errors still there
 > some changes needs to be made to go with Persistence

Suggestion:
I'm not sure what we selected on as our workflow but I think a development branch is needed since we have errors in our main branch right now and this might hinder us on launching applications in the future. But well it's good for now since we're still feeling things out.